Add prevention of published post status changes
This commit introduces new functionality through some new methods
and slight changes to a couple others. Also fixes a PHP warning
when a non-admin tries to Quick Edit a protected post.

The plugin originally used the publish_posts and publish_pages
capabilities as a way to stop non-admins changing the status of
a protected published post to a Draft for example but this
didn't appear to work correctly and also caused a PHP warning
described above by removing the "Published" option from the
post status select box in Quick Edit.

Now we hook into wp_insert_post_data for protected posts and
revert any changes by a non-admin to the post's status, privacy and password.
We also revert changes to the published date if it's changed to
a future date as this would change the post status to future.

If any values are reverted the user will be presented with a
WordPress style admin error notice explaining that they cannot
change the status of the post.

Also introduces the get_admin_cap() which returns the filtered
admin capability required to bypass restrictions.
The method also serves as a callback to the option_page_capability
hook so the old method has been removed.