Fix bug which stopped locked and protected post arrays being filtered.
The filter cap method was being called and therefore calling
load_options() before the arrays had a chance to be filtered
by a theme's functions.php file. I guess this is because WordPress
checks for the 'unfiltered_html' capability before loading the theme.

Fixed by moving the conditional which checks if it's a capability
we're interested in to the top of the filter_cap method.