Remove is_admin or CLI conditional for loading plugin
Gutenberg uses the WP REST API to make requests which check
whether the user has permission to edit or delete posts. As we
were only loading the plugin if is_admin was true or if WP_CLI was
defined and true, and neither of those is true for REST requests,
the plugin was not running and allowing users to edit and delete
posts which were protected or locked.